# challenge-02---elecciones-2023-112866-JulianMartinez/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Parcial 2 - Elecciones 2023'","description":"'Parcial 2 - Elecciones 2023'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1'"},"servers":[{"url":"http://localhost:8081","description":"'Parcial 2 - Elecciones 2023'"}],"paths":{"/secciones":{"get":{"tags":["seccion-controller"],"operationId":"getSecciones","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"seccion_id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Seccion"}}}}}}}},"/resultados":{"get":{"tags":["resultados-controller"],"operationId":"getResultados","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"seccion_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultadoElecciones"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/distrito":{"get":{"tags":["distrito-controller"],"operationId":"getDistritosByName","parameters":[{"name":"nombre","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Distrito"}}}}}}}},"/cargos":{"get":{"tags":["cargo-controller"],"operationId":"getCargos","parameters":[{"name":"distrito_id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CargosDto"}}}}}}}},"components":{"schemas":{"Seccion":{"type":"object","properties":{"seccionId":{"type":"integer","format":"int64"},"seccionNombre":{"type":"string"},"distritoId":{"type":"integer","format":"int64"}}},"ResultadoDto":{"type":"object","properties":{"orden":{"type":"integer","format":"int32"},"nombre":{"type":"string"},"votos":{"type":"integer","format":"int32"},"porcentaje":{"type":"number","format":"double"}}},"ResultadoElecciones":{"type":"object","properties":{"distrito":{"type":"string"},"seccion":{"type":"string"},"resultados":{"type":"array","items":{"$ref":"#/components/schemas/ResultadoDto"}}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"Distrito":{"type":"object","properties":{"distritoId":{"type":"integer","format":"int64"},"distritoNombre":{"type":"string"}}},"Cargo":{"type":"object","properties":{"cargoId":{"type":"integer","format":"int64"},"cargoNombre":{"type":"string"},"distritoId":{"type":"integer","format":"int64"}}},"CargosDto":{"type":"object","properties":{"distrito":{"$ref":"#/components/schemas/Distrito"},"cargos":{"type":"array","items":{"$ref":"#/components/schemas/Cargo"}}}}}}}