# challenge-02---elecciones-2023-112866-JulianMartinez/src/main/java/ar/edu/utn/frc/tup/lc/iv/controllers/DistritoController.java
package ar.edu.utn.frc.tup.lc.iv.controllers;

import ar.edu.utn.frc.tup.lc.iv.dtos.common.DistritoDto;
import ar.edu.utn.frc.tup.lc.iv.models.Distrito;
import ar.edu.utn.frc.tup.lc.iv.services.ElectionsService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
public class DistritoController {

    @Autowired
    private ElectionsService electionsService;

    @GetMapping("/distritos")
    public ResponseEntity<List<DistritoDto>> getDistritosByName(@RequestParam(required = false) String nombre){
        try {
            return new ResponseEntity<>(electionsService.getDistritos(nombre), HttpStatus.OK);
        } catch (Exception e){
            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
        }
    }
}