# challenge-02---elecciones-2023-112866-JulianMartinez/src/main/java/ar/edu/utn/frc/tup/lc/iv/dtos/common/ResultadoElecciones.java
package ar.edu.utn.frc.tup.lc.iv.dtos.common;


import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class ResultadoElecciones {
    private String distrito;
    private String seccion;
    private List<ResultadoDto> resultados;
}